+++ src/Plugin.php
@@ -644,6 +644,8 @@ class Plugin extends BasePlugin
             Plugin::getInstance()->getCarts()->purgeIncompleteCarts();
             // Deletes customers that are not related to any cart/order or user
             Plugin::getInstance()->getCustomers()->purgeOrphanedCustomers();
+            // Deletes addresses that are not related to customers, carts or orders
+            Plugin::getInstance()->getAddresses()->purgeOrphanedAddresses();
         });
     }