+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+- Added purging of orphaned addresses to garbage collection.
+- Added `craft\commerce\services\Addresses::purgeOrphanedAddresses`.
+
 ### Fixed
 - Fixed a bug where Commerce would cause Craft to crash when hosted on Windows machines. ([#1247](https://github.com/craftcms/commerce/issues/1247))
 - Fixed a bug that could occur when editing a completed order causing the Date Ordered date to shift. ([#1246](https://github.com/craftcms/commerce/issues/1246))